+++ packages/charts/react-charts/stories/src/SankeyChart/SankeyChartRebalance.stories.tsx
@@ -374,9 +374,9 @@ export const SankeyChartRebalance: React.FunctionComponent<{}> = (): JSXElement 
           onChange={_onDataSourceChange}
         />
         <label>change Width:</label>
-        <input type="range" value={width} min={400} max={1600} onChange={_onWidthChange} />
+        <input type="range" id="changeWidth_Rebalance" value={width} min={400} max={1600} onChange={_onWidthChange} />
         <label>change Height:</label>
-        <input type="range" value={height} min={312} max={400} onChange={_onHeightChange} />
+        <input type="range" id="changeHeight_Rebalance" value={height} min={312} max={400} onChange={_onHeightChange} />
       </div>
       <div style={rootStyle}>
         <SankeyChart